Update version of normalize dependency
Having v3.0.1 as dependency for normalize causes an extra
folder in the bower_components when the a newer version of
normalize is being downloaded manually. The manually
downloaded normalize would be in "inuit-normalize.css",
whereas the automatic downloaded normalize via this packages
bower.json would be in "inuit-normalize". That leads to two
separate folders with the same code and just a different folder
name. Setting the version dependency in this package for
normalize to v3.0.3 would prevent this minor issue.